Converted the majority of endpoints straight to ESI. Many work!
Known issues:
- Update API key creates a new one
- No employment history fetched
- Citadel authentication broken
- Wallet balance not fetched
- Corp/alliance/faction names not yet routed through EveIDToName or affliation endpoints

+++ src/EVEMon.Common/Serialization/Esi/EsiPlanetaryContentsListItem.cs
@@ -9,6 +9,6 @@ namespace EVEMon.Common.Serialization.Esi
         public int TypeID { get; set; }
 
         [DataMember(Name = "amount")]
-        public long Amount { get; set; }
+        public int Amount { get; set; }
     }
 }
